moved logout code to the settings page
